% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_build_plot.r
\name{add.category.relations}
\alias{add.category.relations}
\title{Add category relations to a soc.ca plot}
\usage{
add.category.relations(
  r,
  cat.rel = filter(get.category.relations(r, dim = dim), valid.correlation),
  dim = c(1, 2),
  mapping = aes(alpha = pem),
  ...
)
}
\arguments{
\item{r}{a soc.mca result object}

\item{cat.rel}{a tibble with category relations like the one produced by \link{get.category.relations}}

\item{dim}{plotting dimensions}

\item{mapping}{aesthetics passed on to geom_segment}

\item{...}{further arguments are passed on to \link[ggplot2]{geom_segment}}
}
\description{
Add category relations to a soc.ca plot
}
