% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_build_plot.r
\name{add.density}
\alias{add.density}
\title{Add a layer with density curves to an mca map.}
\usage{
add.density(
  object,
  dim = c(1, 2),
  ind = extract_ind(object, dim),
  mapping = aes(),
  ...
)
}
\arguments{
\item{object}{a soc.mca result object}

\item{dim}{a numeric vector with the dimensions for the plane}

\item{ind}{a data.frame with coordinates of cases as produced by \link{extract_ind}. This controls the points that are used for the density curves.}

\item{mapping}{a call to \link[ggplot2]{aes} from the ggplot2 package. Here you can map aesthetics to variables such as color, fill, alpha, size and linetype.}

\item{...}{further arguments are passed onto \link[ggplot2]{geom_density_2d}}
}
\description{
Add a layer with density curves to an mca map.
}
