\name{add.to.label}
\alias{add.to.label}
\title{Add to label}
\usage{
  add.to.label(object, text = " (n:")
}
\arguments{
  \item{object}{is a soc.ca object}

  \item{text}{is the prefix used to construct the added
  text.}
}
\value{
  a soc.ca object with altered labels in names.mod and
  names.sup.
}
\description{
  Adds values to the end of the label of each modality.
}

