% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_description.r
\name{breakdown.variance}
\alias{breakdown.variance}
\title{Title}
\usage{
breakdown.variance(object, dim = 1:3, variable)
}
\arguments{
\item{object}{is a soc.ca class object}

\item{dim}{the dimensions in the order they are to be plotted. The first 
number defines the horizontal axis and the second number defines the 
vertical axis.}

\item{variable}{a factor in the same length and order as the active variables}
}
\value{
a matrix
}
\description{
Title
}
\examples{
example(soc.ca)
breakdown.variance(result, dim = 1:3, variable = sup$Gender)
}
