% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_description.r
\name{contribution.headings}
\alias{contribution.headings}
\title{Contribution for headings}
\usage{
contribution.headings(object, dim = 1:3)
}
\arguments{
\item{object}{a \link{soc.ca} object}

\item{dim}{the included dimensions}
}
\value{
a matrix with contribution values for each heading per dimension
}
\description{
Contribution.headings sums the contribution of the active categories according to the heading of their variables.
}
\examples{


}
