% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ellipse.r
\name{ellipses}
\alias{ellipses}
\title{Calculate concentraion ellipses}
\usage{
ellipses(object, var, dim = c(1, 2), kappa = 2, npoints = 1000)
}
\arguments{
\item{object}{a soc.mca result object}

\item{var}{a factor}

\item{dim}{numerical vector with the dimension}

\item{kappa}{the kappa value}

\item{npoints}{number of points}
}
\description{
Calculate concentraion ellipses
}
\examples{
example(soc.mca)
ellipses(result, active[,1])
}
