\name{export}
\alias{export}
\title{Export results from soc.ca}
\usage{
export(object, file = "export.csv", dim = 1:5)
}
\arguments{
  \item{object}{is a soc.ca class object}

  \item{dim}{is the dimensions to be exported}

  \item{file}{is the path and name of the .csv values are
  to be exported to}
}
\value{
A .csv file with various values in UTF-8 encoding
}
\description{
Export objects from the soc.ca package to csv files.
}
\seealso{
\link{soc.mca}, \link{contribution}
}

