% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_build_plot.r
\name{extract_cases}
\alias{extract_cases}
\alias{extract_ind}
\title{Extract individuals}
\usage{
extract_cases(result, dim = 1:3)

extract_ind(result, dim = 1:3)
}
\arguments{
\item{result}{a soc.ca object or a PCA}

\item{dim}{the dimensions}
}
\value{
a data.frame with coordinates and frequences
}
\description{
Extract individuals
}
\examples{
example(soc.mca)
extract_cases(result)
}
