% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_build_plot.r
\name{extract_cats}
\alias{extract_cats}
\alias{extract_mod}
\title{Extract coordinates for the categories from an soc.mca}
\usage{
extract_cats(result, dim = 1:3)

extract_mod(result, dim = 1:3)
}
\arguments{
\item{result}{a soc.mca object or a PCA}

\item{dim}{the dimensions}
}
\value{
a data.frame with coordinates and frequences
}
\description{
Extract coordinates for the categories from an soc.mca
}
\examples{
example(soc.mca)
extract_cats(result)
}
