% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_build_plot.r
\name{extract_ind}
\alias{extract_ind}
\title{Extract individuals}
\usage{
extract_ind(result, dim = 1:2)
}
\arguments{
\item{result}{a soc.ca object}

\item{dim}{the dimensions}
}
\value{
a data.frame with coordinates and frequences
}
\description{
Extract individuals
}
\examples{
example(soc.mca)
extract_ind(result)
}
