% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_description.r
\name{headings}
\alias{headings}
\title{Calculate contributions per heading}
\usage{
headings(object, dim = 1:5)
}
\arguments{
\item{object}{a soc.ca object with headings}

\item{dim}{a numeric vector with the dimensions}
}
\value{
a matrix
}
\description{
Calculate contributions per heading
}
\examples{
data(taste)
active.headings <- list()
active.headings$Consumption <- taste[, c("TV", "Film", "Art", "Eat")
active.headings$Background  <- taste[, c("Gender", "Age", "Income")]
result.headings <- soc.mca(active.headings, active.is.list = TRUE)
}
