% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_shiny.r
\name{ind.explorer}
\alias{ind.explorer}
\title{Explore the cloud of individuals}
\usage{
ind.explorer(object, active, sup = NULL)
}
\arguments{
\item{object}{a a soc.ca class object as created by \link{soc.mca} and 
\link{soc.csa}}

\item{active}{Defines the active modalities in a data.frame with rows of individuals and columns of factors, without NA's'}

\item{sup}{Defines the supplementary modalities in a data.frame with rows of individuals and columns of factors, without NA's}
}
\value{
an html application
}
\description{
Explore the cloud of individuals
}
\examples{
\dontrun{
examples(soc.mca)
ind.explorer(result, active, sup)
}
}

