% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_tools.r
\name{indicator.to.long}
\alias{indicator.to.long}
\title{Pivot the indicator matrix from an MCA to long format}
\usage{
indicator.to.long(r)
}
\arguments{
\item{r}{an result object of class soc.mca}
}
\value{
a tibble in long format
}
\description{
Sometimes we want the indicator matrix from an mca in long format and this function delivers just that. The results contain both active and passive categories.
}
\examples{
example(soc.mca)
result |> indicator.to.long() 
}
