% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_build_plot.r
\name{map.ca.base}
\alias{map.ca.base}
\title{Create the base of a soc.ca map}
\usage{
map.ca.base(
  up = NULL,
  down = NULL,
  right = NULL,
  left = NULL,
  base_size = 15,
  ...
)
}
\arguments{
\item{up}{the name of + pole on the vertical axis - "North"}

\item{down}{the name of the - pole on the vertical axis - "South"}

\item{right}{the name of the + pole on horizontal axis - "East"}

\item{left}{the name of the - pole on the horizontal axis - "West"}

\item{base_size}{controls the text size of themed labels}

\item{...}{further arguments are passed onto ggplot()}
}
\value{
a ggplot2 object
}
\description{
Create the base of a soc.ca map
}
