\name{map.modal}
\alias{map.modal}
\title{Plot all points}
\usage{
  map.modal(object, dim = c(1, 2), map.title = "all",
    labelx = NULL, labely = NULL, point.label = TRUE,
    point.shape = "variable", point.size = "freq",
    text.size = 3.3, colour = "black")
}
\description{
  Plot all points
}

