% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.r
\docType{data}
\name{pe13}
\alias{pe13}
\title{The Field of the Danish Power Elite}
\description{
This dataset was used to construct a field of the Danish Power Elite from 2013
}
\author{
Jacob Lunding, Anton Grau Larsen and Christoph Ellersgaard
}
\keyword{data}
