% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_description.r
\name{print.soc.mca}
\alias{print.soc.mca}
\title{Print soc.ca objects}
\usage{
\method{print}{soc.mca}(x, ...)
}
\arguments{
\item{x}{is a soc.ca class object}

\item{...}{further arguments are ignored}
}
\value{
Active dimensions is the number of dimensions remaining after the
  reduction of the dimensionality of the analysis.

Active modalities is the number of modalities that are not set as
  passive.

Share of passive mass is the percentage of the total mass that is
  represented by the passive modalities.

The values represented in the scree plot are the adjusted inertias,
  see \link{variance}

The active variables are represented with their number of active
  modalities and their share of the total variance/inertia.
}
\description{
Prints commonly used measures used in the analysis of multiple correspondence
analysis
}
\examples{
example(soc.ca)
print(result)
}
\seealso{
\link{soc.mca}, \link{contribution}
}
