% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_tools.r
\name{randomize.mca}
\alias{randomize.mca}
\title{Create a randomized mca on the basis of an existing mca}
\usage{
randomize.mca(r, replace = FALSE)
}
\arguments{
\item{r}{a result object from soc.mca}

\item{replace}{}
}
\value{
a soc.mca object
}
\description{
We sample from each of the active variables independently removing the
original correlations but retaining the frequencies of the categories. This
function is useful to see the extent to which the mca solution reflects the
correlations between variables or the frequency distribution between the
active categories. Passive categories are inherited from the original analysis.
}
\examples{
example(soc.mca)
randomize.mca(result)
}
