% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_build_plot.r
\name{rename_aes}
\alias{rename_aes}
\title{rename_aes
 
We import this function from ggplot2 because it is not exported.}
\usage{
rename_aes(x)
}
\arguments{
\item{x}{an aes}
}
\description{
rename_aes
 
We import this function from ggplot2 because it is not exported.
}
