% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_description.r
\name{supplementary.categories}
\alias{supplementary.categories}
\title{Supplementary coordinates for a data.frame of factors}
\usage{
supplementary.categories(object, sup, dim = 1:2)
}
\arguments{
\item{object}{a soc.mca result object}

\item{sup}{a data.frame of factors or an indicator matrix}

\item{dim}{a numeric vector with the two dimensions calculated}
}
\value{
a data.frame with coordinates and labels
}
\description{
Calculate the average coordinates in the category cloud of a soc.mca analysis.
}
\examples{
example(soc.mca)
supplementary.categories(result, sup)
}
