\docType{data}
\name{taste}
\alias{taste}
\title{Taste dataset
The taste example dataset used by Le Roux & Rouanet(2010):}
\value{
  The variables included in the dataset: \item{Preferred TV
  program}{(8 categories): news, comedy, police, nature,
  sport, films, drama, soap operas} \item{Preferred
  Film}{(8 categories): action, comedy, costume drama,
  documentary, horror, musical, romance, SciFi}
  \item{Preferred type of Art}{(7 categories): performance,
  landscape, renaissance, still life, portrait, modern,
  impressionsism} \item{Preferred place to Eat out}{(6
  categories): fish & chips, pub, Indian restuarant,
  Italian restaurant, French restaurant, steak house}
}
\description{
  Taste dataset The taste example dataset used by Le Roux &
  Rouanet(2010):
}
\author{
  Brigitte Le Roux
}
\references{
  Le Roux, Brigitte, Henry Rouanet, Mike Savage, og Alan
  Warde. 2008. "Class and Cultural Division in the UK".
  Sociology 42(6):1049-1071.

  Le Roux, B., og H. Rouanet. 2010. Multiple correspondence
  analysis. Thousand Oaks: Sage.
}
\keyword{data}
\keyword{datasets}

