% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_tools.r
\name{to.MCA}
\alias{to.MCA}
\title{Convert to MCA class from FactoMineR}
\usage{
to.MCA(object, active, dim = 1:5)
}
\arguments{
\item{object}{is a soc.ca object}

\item{active}{the active variables}

\item{dim}{a numeric vector}
}
\value{

}
\description{
Convert to MCA class from FactoMineR
}
