% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_description.r
\name{variance}
\alias{variance}
\title{Variance table}
\usage{
variance(object, dim = NULL)
}
\arguments{
\item{object}{is a soc.ca object}

\item{dim}{is the included dimensions, if set to NULL, then only the
dimensions explaining approx. more than 0.90 of the adjusted variance are included}
}
\value{
If assigned variance returns a matrix version of the table
  of variance.
}
\description{
variance returns a table of variance for the selected dimensions.
}
\examples{
example(soc.ca)
variance(result)
variance(result, dim = 1:4)
}
\seealso{
\link{soc.mca}, \link{print.soc.mca}
}
