\name{add.count}
\alias{add.count}
\title{Add a new layer of points on top of an existing plot with output from the cut.min function}
\usage{
  add.count(x, p = map.ind(object), label = TRUE, ...)
}
\arguments{
  \item{x}{a matrix created by the cut.min function}

  \item{p}{is a ggplot object, preferably from one of the
  mapping functions in soc.ca}

  \item{label}{if TRUE the labels of points will be shown}
}
\description{
  Add a new layer of points on top of an existing plot with
  output from the cut.min function
}

