% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions_description.r
\name{average.coord}
\alias{average.coord}
\title{Average coordinates}
\usage{
average.coord(object, x, dim = c(1, 2))
}
\arguments{
\item{object}{is soc.ca result object}

\item{x}{is a variable of the same length and order as the active variables used to construct the soc.ca object}

\item{dim}{is the two dimensions used}
}
\value{
a matrix with the mean points and frequencies of the given variable
}
\description{
Find the average coordinates for each category in a variable on two dimensions.
}
\examples{
example(soc.ca)
average.coord(result, sup$Income)
}

