% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ellipse.r
\name{ellipses}
\alias{ellipses}
\title{Calculate concentraion ellipses}
\usage{
ellipses(object, var, dim = c(1, 2), kappa = 2, npoints = 1000)
}
\arguments{
\item{object}{}

\item{var}{}

\item{dim}{}

\item{kappa}{}

\item{npoints}{}
}
\value{

}
\description{
Calculate concentraion ellipses
}
\examples{
example(soc.mca)
ellipses(result, active[,1])
}
