% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_tools.r
\name{export}
\alias{export}
\title{Export results from soc.ca}
\usage{
export(object, file = "export.csv", dim = 1:5)
}
\arguments{
\item{object}{is a soc.ca class object}

\item{file}{is the path and name of the .csv values are to be exported to}

\item{dim}{is the dimensions to be exported}
}
\value{
A .csv file with various values in UTF-8 encoding
}
\description{
Export objects from the soc.ca package to csv files.
}
\seealso{
\link{soc.mca}, \link{contribution}
}
