% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_build_plot.r
\name{extract_ind}
\alias{extract_ind}
\title{Extract individuals}
\usage{
extract_ind(result, dim = 1:2)
}
\arguments{
\item{result}{}

\item{dim}{}
}
\value{

}
\description{
Extract individuals
}
