% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_build_plot.r
\name{extract_sup}
\alias{extract_sup}
\title{Extract supplementary categories}
\usage{
extract_sup(result, dim = 1:2)
}
\arguments{
\item{result}{}

\item{dim}{}
}
\value{

}
\description{
Extract supplementary categories
}
