% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions_tools.r
\name{invert}
\alias{invert}
\title{Invert the direction of coordinates}
\usage{
invert(x, dim = 1)
}
\arguments{
\item{x}{is a soc.ca object}

\item{dim}{is the dimensions to be inverted}
}
\value{
a soc.ca object with inverted coordinates on the specified dimensions
}
\description{
Invert one or more axes of a correspondence analysis. The principal coordinates of the analysis are multiplied by -1.
}
\details{
This is a convieniency function as you would have to modify coord.mod, coord.ind and coord.sup in the soc.ca object.
}
\examples{
example(soc.ca)
inverted.result  <- invert(result, 1:2)
result$coord.ind[1, 1:2]
inverted.result$coord.ind[1, 1:2]
}
\seealso{
\link{soc.mca}, \link{add.to.label}
}

