% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_build_plot.r
\name{map.ca.base}
\alias{map.ca.base}
\title{Create the base of a soc.ca map}
\usage{
map.ca.base(up = NULL, down = NULL, right = NULL, left = NULL, ...)
}
\arguments{
\item{up}{}

\item{down}{}

\item{right}{}

\item{left}{}

\item{...}{}
}
\value{

}
\description{
Create the base of a soc.ca map
}
