% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_triads.r
\name{mca.triads}
\alias{mca.triads}
\title{Title}
\usage{
mca.triads(l.mca, l.triads, dim = c(1, 2), fix.mca = 1)
}
\arguments{
\item{l.mca}{a list of soc.mca objects}

\item{l.triads}{}

\item{dim}{}

\item{fix.mca}{}
}
\value{

}
\description{
Title
}
