\docType{package}
\name{soc.ca}
\alias{soc.ca}
\alias{soc.ca-package}
\title{Soc.ca a package for specific correspondence analysis}
\description{
  This package is optimized to the needs of sociologists
  producing specific and class specific multiple
  correspondence analysis on survey-like data. Soc.ca is
  optimized to only give the most essential statistical
  output values sorted so as to help in analysis. Seperate
  functions exists for near publication-ready plots and
  tables.
}
\details{
  We are in debt to the great work of others, especially
  Brigitte Le Roux and Henry Rouanet for the mathematical
  definitions of the method and their examples. Furthermore
  this package was initially based on code from the ca
  package written by Michael Greenacre and Oleg Nenadic.

  If you are looking for features that are absent in
  soc.ca, it may be available in some of these great
  packages for correspondence analysis: \pkg{ca},
  \pkg{anacor} and \pkg{FactoMineR}.
}
\references{
  Le Roux, B., og H. Rouanet. 2010. Multiple correspondence
  analysis. Thousand Oaks: Sage.
}

